+++ R/SimResults.R
@@ -2,10 +2,11 @@
 #'
 #' If \code{\link{runSimulation}} was passed the flag \code{save_results = TRUE} then the
 #' row results corresponding to the \code{design} object will be stored to a suitable
-#' sub-directory as individual \code{.rds} files. While users could use \code{\link{readRDS}} directly
+#' sub-directory as individual \code{.rds} or \code{qs2} files.
+#' While users could use \code{\link{readRDS}} and \code{qs2} directly
 #' to read these files in themselves, this convenience function will read the desired rows in
 #' automatically given the returned object
-#' from the simulation. Can be used to read in 1 or more \code{.rds} files at once (if more than 1 file
+#' from the simulation. Can be used to read in 1 or more files at once (if more than 1 file
 #' is read in then the result will be stored in a list).
 #'
 #' @param obj object returned from \code{\link{runSimulation}} where \code{save_results = TRUE}
@@ -56,7 +57,7 @@
 #'
 #' @author Phil Chalmers \email{rphilip.chalmers@@gmail.com}
 #'
-#' @seealso \code{\link{descript}}
+#' @seealso \code{\link{descript}}, \code{\link{SimRead}}
 #'
 #' @examples
 #'