+++ R/RobbinsMonro.R
@@ -1,6 +1,6 @@
 #' Robbins-Monro (1951) stochastic root-finding algorithm
 #'
-#' Function performs stochastic root solving for the provided \code{f(x)}
+#' Performs stochastic root solving for the provided \code{f(x)}
 #' using the Robbins-Monro (1951) algorithm. Differs from deterministic
 #' cousins such as \code{\link{uniroot}} in that \code{f} may contain stochastic error
 #' components, where the root is obtained through the running average method